Moved crosshair creation to util
